genImages/pyproject.toml
[project]
name = "mcp-image-gen"
version = "0.1.0"
description = "MCP server for AI image generation via ComfyUI + FLUX.1"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.9.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "pillow>=11.0.0",
    "aiofiles>=24.1.0",
    "python-multipart>=0.0.20",
]

[project.scripts]
mcp-image-gen = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
